refactor(frontends/basic): introduce statement sequencer helper
* Move statement sequencing into a dedicated StatementSequencer helper shared by parser routines.

* Update Parser statements to consume the helper interface and remove StatementContext.

* Expand parser tests to cover multiline, colon, and labeled statement scenarios.

- [x] Built on macOS/Linux (Debug).

- [x] ctest passed.

- [x] Added/updated tests.

- [x] Code formatted; no new warnings.

- [x] No spec/API deviation (or ADR attached).

- [x] Docs updated if behavior/API changed.

- [x] Updated docs/comments per policy.

- [x] Added/updated file headers.
