+++ src/frontends/basic/Parser.hpp
@@ -9,6 +9,7 @@
 #include "frontends/basic/AST.hpp"
 #include "frontends/basic/DiagnosticEmitter.hpp"
 #include "frontends/basic/Lexer.hpp"
+#include "frontends/basic/StatementSequencer.hpp"
 #include <array>
 #include <functional>
 #include <initializer_list>
@@ -35,100 +36,24 @@ class Parser
     std::unique_ptr<Program> parseProgram();
 
   private:
-    /// @brief Helper that manages statement parsing loops and shared token patterns.
-    class StatementContext
-    {
-      public:
-        /// @brief Aggregated information about a terminating keyword.
-        struct TerminatorInfo
-        {
-            int line = 0; ///< Optional line number that preceded the terminator.
-            il::support::SourceLoc
-                loc{}; ///< Source location where the terminator keyword appeared.
-        };
+    friend class StatementSequencer;
 
-        /// @brief Classification of the last separator consumed by the context.
-        enum class SeparatorKind
-        {
-            None,      ///< No separator consumed since the previous statement.
-            Colon,     ///< A colon separated the previous and next statements.
-            LineBreak, ///< A line break separated the previous and next statements.
-        };
-
-        using TerminatorPredicate =
-            std::function<bool(int)>; ///< Predicate identifying terminator tokens.
-        using TerminatorConsumer = std::function<void(
-            int, TerminatorInfo &)>; ///< Callback consuming the terminator tokens.
-
-        /// @brief Construct a context bound to @p parser.
-        /// @param parser Owning parser providing token accessors.
-        explicit StatementContext(Parser &parser);
-
-        /// @brief Consume a single leading colon or end-of-line if present.
-        void skipLeadingSeparator();
-
-        /// @brief Consume all consecutive end-of-line tokens.
-        /// @return True when at least one line break was consumed.
-        bool skipLineBreaks();
-
-        /// @brief Consume a colon or end-of-line token if immediately present.
-        void skipStatementSeparator();
-
-        /// @brief Invoke @p fn with an optional numeric line label.
-        /// @param fn Callback receiving the parsed line number (zero when absent).
-        void withOptionalLineNumber(const std::function<void(int)> &fn);
-
-        /// @brief Remember a line label that should seed the next statement.
-        /// @param line Parsed line number to reuse on the next call.
-        void stashPendingLine(int line);
-
-        /// @brief Report which separator most recently separated statements.
-        /// @return Classification of the most recent separator.
-        SeparatorKind lastSeparator() const;
-
-        /// @brief Parse statements until @p isTerminator matches and populate @p dst.
-        /// @param isTerminator Predicate determining when the body ends.
-        /// @param onTerminator Callback consuming the terminator once detected.
-        /// @param dst Destination vector receiving parsed statements.
-        /// @return Line/location metadata of the terminating keyword.
-        TerminatorInfo consumeStatementBody(const TerminatorPredicate &isTerminator,
-                                            const TerminatorConsumer &onTerminator,
-                                            std::vector<StmtPtr> &dst);
-
-        /// @brief Parse statements until @p terminator is encountered.
-        /// @param terminator Token that terminates the body.
-        /// @param dst Destination vector receiving parsed statements.
-        /// @return Line/location metadata of the terminator keyword.
-        TerminatorInfo consumeStatementBody(TokenKind terminator, std::vector<StmtPtr> &dst);
-
-      private:
-        Parser &parser_;       ///< Parent parser that owns the token stream.
-        int pendingLine_ = -1; ///< Deferred line label for the next statement.
-        SeparatorKind lastSeparator_ =
-            SeparatorKind::LineBreak; ///< Classification of the last separator consumed.
-    };
-
-    /// @brief Create a statement context bound to this parser instance.
-    /// @return StatementContext referencing the parser's token stream.
-    StatementContext statementContext();
-
-    /// @brief Parse the next logical statement line into a single AST node.
-    /// @param ctx Statement context that manages separators and line labels.
-    /// @return Either a single statement or a StmtList when multiple statements share a line.
-    StmtPtr parseStatementLine(StatementContext &ctx);
+    /// @brief Create a statement sequencer bound to this parser instance.
+    /// @return StatementSequencer referencing the parser's token stream.
+    StatementSequencer statementSequencer();
 
     /// @brief Consume optional line labels that follow a line break.
-    /// @param ctx Statement context providing newline skipping helpers.
+    /// @param ctx Statement sequencer providing newline skipping helpers.
     /// @param followerKinds When non-empty, only consume the label when the
     ///        subsequent token is one of the specified kinds.
-    void skipOptionalLineLabelAfterBreak(StatementContext &ctx,
+    void skipOptionalLineLabelAfterBreak(StatementSequencer &ctx,
                                          std::initializer_list<TokenKind> followerKinds = {});
 
     /// @brief Parse the body of a single IF-related branch.
     /// @param line Line number propagated to the branch statement.
-    /// @param ctx Statement context for separator management.
+    /// @param ctx Statement sequencer for separator management.
     /// @return Parsed statement belonging to the branch body.
-    StmtPtr parseIfBranchBody(int line, StatementContext &ctx);
+    StmtPtr parseIfBranchBody(int line, StatementSequencer &ctx);
 
     mutable Lexer lexer_;                    ///< Provides tokens from the source buffer.
     mutable std::vector<Token> tokens_;      ///< Lookahead token buffer.
